[#30] fix: 그룹 서비스 역할 분리
- 기존 그룹 서비스에서 그룹의 인증과 관리하는 역할을 하였습니다.
- 수정 후, 그룹 서비스를 그룹 인증 서비스와 그룹 관리 서비스로 역할을 분리하였습니다.
